+++ include/policy/sycl_policy_handler.h
@@ -54,7 +54,7 @@ class PolicyHandler<codeplay_policy> {
         computeUnits_(codeplay_policy::get_num_compute_units(q)) {}
 
   template <typename element_t>
-  element_t *allocate(size_t num_elements) const;
+  element_t *allocate(int num_elements) const;
 
   template <typename element_t>
   void deallocate(element_t *p) const;
@@ -124,7 +124,7 @@ class PolicyHandler<codeplay_policy> {
 
   template <typename element_t>
   typename policy_t::event_t copy_to_device(const element_t *src,
-                                            element_t *dst, size_t size = 0);
+                                            element_t *dst, int size = 0);
   /*  @brief Copying the data back to device
     @tparam element_t is the type of the data
     @param src is the host pointer we want to copy from.
@@ -135,7 +135,7 @@ class PolicyHandler<codeplay_policy> {
   template <typename element_t>
   typename policy_t::event_t copy_to_device(
       const element_t *src, BufferIterator<element_t, policy_t> dst,
-      size_t size = 0);
+      int size = 0);
   /*  @brief Copying the data back to device
       @tparam element_t is the type of the data
       @param src is the device pointer we want to copy from.
@@ -145,16 +145,16 @@ class PolicyHandler<codeplay_policy> {
 
   template <typename element_t>
   typename policy_t::event_t copy_to_host(element_t *src, element_t *dst,
-                                          size_t size = 0);
+                                          int size = 0);
 
   template <typename element_t>
   typename policy_t::event_t copy_to_host(
-      BufferIterator<element_t, policy_t> src, element_t *dst, size_t size = 0);
+      BufferIterator<element_t, policy_t> src, element_t *dst, int size = 0);
 
   template<typename element_t>
   typename policy_t::event_t fill(BufferIterator<element_t, policy_t> buf,
                                   element_t value = element_t{0},
-                                  size_t size = 0);
+                                  int size = 0);
 
   inline const policy_t::device_type get_device_type() const {
     return selectedDeviceType_;
